refactor: changed navigation to be service
