+++ src/services/NotificationModule.ts
@@ -1,4 +1,3 @@
-import { useEffect } from 'react';
 import * as Notifications from 'expo-notifications';
 
 export default function NotificationModule() {
