+++ src/utils/navigate.ts
@@ -1,4 +1,3 @@
-// services/NavigationService.ts
 import { createNavigationContainerRef, ParamListBase } from '@react-navigation/native';
 
 export const navigationRef = createNavigationContainerRef<ParamListBase>();
